+++ releases/dogwood/3/fun/entrypoint.sh
@@ -5,4 +5,8 @@
 
 # Activate user's virtualenv
 source /edx/app/edxapp/venv/bin/activate
+
+# Override default root_urls for the lms app
+ln -sf /config/lms/root_urls.py /edx/app/edxapp/edx-platform/lms/
+
 exec "$@"